Add static analysis and code coverage to CI
- clang-tidy (bugprone/cert/clang-analyzer/misc/performance/portability/
  most of readability) and cppcheck, each with justified suppressions in
  .clang-tidy for checks that are false positives on this project's target
  platforms or fight an established convention, documented inline.
- Two real findings fixed directly: a dead variable initializer in
  osal_queue_send/receive, and an int loop index in can_protocol.c's
  cell-voltage pack/unpack implicitly widened to a pointer offset (size_t
  is correct regardless of the current tiny bound). One cppcheck finding
  suppressed instead: task_can.c's redundant-looking "&& i < 4" is a real
  guard against overflowing the CAN wire format's fixed 4-slot array if
  BMS_CELL_COUNT ever grows, not dead code.
- make coverage: per-file --coverage build + lcov report, scoped to src/,
  no enforced threshold -- it's there to make the existing "what's and
  isn't unit tested" philosophy numerically visible, not to gate on it.
- Verified against the actual CI tool versions (Ubuntu's clang-tidy
  differs from a locally-installed one and enabled an extra check
  suggesting C11 Annex K functions neither glibc nor macOS implement),
  same lesson as the earlier _POSIX_C_SOURCE build fix.

+++ src/can/can_protocol.c
@@ -33,14 +33,14 @@ void can_protocol_pack_cell_voltages(const cell_voltages_t *cv, can_frame_t *out
     memset(out, 0, sizeof(*out));
     out->id = CAN_ID_CELL_VOLTAGES;
     out->dlc = 8;
-    for (int i = 0; i < 4; i++) {
+    for (size_t i = 0; i < 4; i++) {
         put_u16(&out->data[i * 2], cv->cell_mv[i]);
     }
 }
 
 int can_protocol_unpack_cell_voltages(const can_frame_t *in, cell_voltages_t *out) {
     if (in->id != CAN_ID_CELL_VOLTAGES || in->dlc < 8) return -1;
-    for (int i = 0; i < 4; i++) {
+    for (size_t i = 0; i < 4; i++) {
         out->cell_mv[i] = get_u16(&in->data[i * 2]);
     }
     return 0;
